First commit
Begining to work with and learn Pester for when I do Powershell. This is some of the first stuf I quickly joted down while watching a tutorial.

""My thourth:""
I'm putting this in it's own branch to split it from the master repo, as this is in no way ready for use in anyway. It's still on the sandbox/playground stage.